% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available.data}
\alias{available.data}
\title{Available data}
\format{
A data frame with 2 rows (datasets) and 5 columns (attributes):
\describe{
\item{filename}{Dataset}
\item{md5}{MD5 checksum}
\item{url}{URL link}
\item{date}{Date updated}
\item{size}{Data size}
}
}
\usage{
available.data
}
\description{
Available data can be loaded using \code{\link[spots]{LoadData}} function.
}
\concept{data}
\keyword{datasets}
