% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{spots-package}
\alias{spots-package}
\alias{_PACKAGE}
\title{spots: Spatial Component Analysis}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The spots package is designed for spatial omics (10x Visium, etc.) data analysis. It performs various statistical analyses and tests, including spatial component analysis (SCA), both global and local spatial statistics, such as univariate and bivariate Moran's I, Getis-Ord Gi* statistics, etc. See Integrated protein and transcriptome high-throughput spatial profiling (2022) <doi:10.1101/2022.03.15.484516> for more details.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/stevexniu/spots/issues}
}

}
\author{
\strong{Maintainer}: X. Steve Niu \email{xin2001@med.cornell.edu} (\href{https://orcid.org/0000-0001-7689-0941}{ORCID})

}
