% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnn.R
\name{VisiumHnn}
\alias{VisiumHnn}
\title{10x Genomics Visium hexagonal nearest neighbor}
\usage{
VisiumHnn(barcodes)
}
\arguments{
\item{barcodes}{10x Visium whitelisted spatial barcodes.}
}
\value{
A hexagonal distance matrix for selected barcodes.
}
\description{
Get hexagonal nearest neighbor distance between 10x Visium spatial barcodes.
}
\examples{
\dontrun{
barcodes <- c("AAACAACGAATAGTTC-1", "AAACTTAATTGCACGC-1")
VisiumHnn(barcodes)
}

}
\seealso{
\code{\link[rspca]{LoadData}}.
}
\concept{hnn}
