% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnn.R
\name{VisiumHnn}
\alias{VisiumHnn}
\title{10x Genomics Visium hexagonal nearest neighbor}
\usage{
VisiumHnn(path, barcodes)
}
\arguments{
\item{path}{Path to downloaded HNN distance for Visium barcodes. See \code{\link[spots]{available.data}} for download url.}

\item{barcodes}{10x Visium whitelisted spatial barcodes.}
}
\value{
A hexagonal distance matrix for selected barcodes.
}
\description{
Get hexagonal nearest neighbor distance between 10x Visium spatial barcodes.
}
\examples{
\dontrun{
barcodes <- c("AAACAACGAATAGTTC-1", "AAACTTAATTGCACGC-1")
VisiumHnn("~/Downloads/", barcodes)
}

}
\seealso{
\code{\link[spots]{available.data}}.
}
\concept{hnn}
